Move model_to_digraph function to utils module
